pyproject.toml
[tool.poetry]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = ["Dimanskr <dimanskr2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pillow = "^11.0.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.51"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
